# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaravelAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/wp_clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Get Clientes From Contact Form",
                "description": "Obtiene todos los clientes del contact form",
                "operationId": "getWPClientes",
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No Autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string",
                                                "example": "Luis Ventocilla"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "lvsmix@gmail.com"
                                            },
                                            "topic": {
                                                "type": "string",
                                                "example": "Consulta"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Prueba"
                                            },
                                            "created": {
                                                "type": "string",
                                                "example": "2020-09-17 03:21:05"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/migrarclientes": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Migrar Clientes",
                "description": "Migrar todos los clientes obtenidos desde el contact form",
                "operationId": "migrarClientes",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Process Succesfull"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Get Clientes",
                "description": "Obtiene todos los clientes migrados",
                "operationId": "getClientes",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string",
                                                "example": "Luis Ventocilla"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "lvsmix@gmail.com"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "INGRESADO"
                                            },
                                            "created": {
                                                "type": "string",
                                                "example": "2020-09-17 03:21:05"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}